src/requirements.txt
pyserial>=3.4
keyboard>=0.13.5
